[FEAT/#5] SpoonyTextStyle 고정되는 디폴트값 추가 및 Theme.kt 안에 있는 SpoonyColors Color.kt 파일로 이동 및 Color Preview 추가
